CCMngr propagates disconnected
